+++ sql/01-schema/02_book_sale_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE book_sale
 (
-    sale_id SERIAL NOT NULL,
-    book_id INT NOT NULL,
+    sale_id BIGINT NOT NULL,
+    book_id BIGINT NOT NULL,
     quantity INT NOT NULL,
     date_sold TIMESTAMP DEFAULT NOW(),
     amount NUMERIC(10,2) NOT NULL
